backend/tag-service/requirements.txt
fastapi
uvicorn
supabase
asyncpg
python-dotenv==1.0.0
pydantic